+++ src/game/server/tf/bot/tf_bot.h
@@ -358,6 +358,8 @@ public:
 		MISSION_SPY,					// maintain teams of spies harassing the enemy
 		MISSION_ENGINEER,				// maintain engineer nests for harassing the enemy
 		MISSION_REPROGRAMMED,			// MvM: robot has been hacked and will do bad things to their team
+
+		MISSION_CHOREO,					// get into position for choreo
 	};
 	#define MISSION_DOESNT_RESET_BEHAVIOR_SYSTEM false
 	void SetMission( MissionType mission, bool resetBehaviorSystem = true );